remove inserting/fetching from submission table
we are only using the post table
also create challenge page
