+++ src/app/challenge/[id].tsx
@@ -1,0 +1,20 @@
+import { Stack } from 'expo-router';
+import ChallengePage from '../../components/Challenge';
+import { StyleSheet, SafeAreaView } from 'react-native';
+
+export default function ChallengePageWrapper() {
+  return (
+    <>
+      <Stack.Screen options={{ headerShown: false }} />
+      <SafeAreaView style={styles.container}>
+        <ChallengePage />
+      </SafeAreaView>
+    </>
+  );
+}
+
+const styles = StyleSheet.create({
+  container: {
+    flex: 1,
+  },
+});
