+++ src/constants/Patrizio.ts
@@ -1,0 +1,1 @@
+export const PATRIZIO_ID = 'b30a8bbd-7131-44c3-a746-8c726e0bb64c';
